CouponDiscountPolicy -> DiscountPolicy changed and moved to domain
