+++ src/main/java/app/foodylab/domain/discount/DiscountPolicy.java
@@ -1,8 +1,9 @@
-package app.foodylab.application.coupon;
+package app.foodylab.domain.discount;
 
+import app.foodylab.application.coupon.Coupon;
 import app.foodylab.domain.coupon.CouponType;
 
-public interface CouponDiscountPolicy {
+public interface DiscountPolicy {
 
     long getCalculatedDiscountPrice(Coupon coupon, long productPrice);
