+++ src/testFixtures/java/app/foodylab/coupon/PriceDiscountPolicySpy.java
@@ -1,7 +1,5 @@
 package app.foodylab.coupon;
 
-import app.foodylab.application.coupon.PriceDiscountPolicy;
-
 public class PriceDiscountPolicySpy extends PriceDiscountPolicy {
 
 }
