+++ src/testFixtures/java/app/foodylab/coupon/RateDiscountPolicySpy.java
@@ -1,7 +1,5 @@
 package app.foodylab.coupon;
 
-import app.foodylab.application.coupon.RateDiscountPolicy;
-
 public class RateDiscountPolicySpy extends RateDiscountPolicy {
 
 }
